Assistant checkpoint: Modifica il tipo del campo prezzo da integer a float
Assistant generated file changes:
- python_server/models/models.py: Cambia il tipo del campo price da integer a float
- python_server/config/settings.py: Modifica il mapping OpenSearch per supportare un prezzo decimale, Modifica il mapping SavedPackages per supportare un prezzo decimale
- .replit: Updated app configuration

---

User prompt:

ancora errore:

+++ python_server/config/settings.py
@@ -86,7 +86,7 @@ MAPPINGS = {
                 "durationDays": {"type": "integer"},
                 "durationNights": {"type": "integer"},
                 "experiences": {"type": "keyword"}, # Assuming experiences is array of strings in the frontend TravelPackage
-                "price": {"type": "integer"},
+                "price": {"type": "float"},
                 "isRecommended": {"type": "boolean"},
                 "categories": {"type": "keyword"}
             }
@@ -130,7 +130,7 @@ MAPPINGS = {
                 "durationDays": {"type": "integer"},
                 "durationNights": {"type": "integer"},
                 "experiences": {"type": "keyword"},  # Assuming experiences is array of strings in the frontend TravelPackage
-                "price": {"type": "integer"},
+                "price": {"type": "float"},
                 "isRecommended": {"type": "boolean"},
                 "categories": {"type": "keyword"}
             }
